requirements.txt
numpy
torch
transformers
Pillow
pytest
jsonschema